dom/nodes/CharacterData-replaceData.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>CharacterData.replaceData</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-replacedata">
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-data">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
// TODO: Non-ASCII strings.
function testNode(create, type) {
  // Step 2.
  test(function() {
    var node = create()
    assert_equals(node.data, "test")

    assert_throws("IndexSizeError", function() { node.replaceData(5, 1, "x") })
    assert_throws("IndexSizeError", function() { node.replaceData(5, 0, "") })
    assert_equals(node.data, "test")
  }, type + ".replaceData() with invalid offset")

  // Step 3.
  test(function() {
    var node = create()
    assert_equals(node.data, "test")

    node.replaceData(2, 10, "yo")
    assert_equals(node.data, "teyo")
  }, type + ".replaceData() with clamped count")

  test(function() {
    var node = create()
    assert_equals(node.data, "test")

    node.replaceData(1, 1, "waddup")
    assert_equals(node.data, "twaddupst")
    node.replaceData(1, 1, "yup")
    assert_equals(node.data, "tyupaddupst")
  }, type + ".replaceData() in the middle")
}

testNode(function() { return document.createTextNode("test") }, "Text")
testNode(function() { return document.createComment("test") }, "Comment")
</script>
